[ADMIN] Re-organize certification centers components
