+++ admin/app/templates/authenticated/certification-centers/get.hbs
@@ -49,6 +49,6 @@
     {{/if}}
   </section>
 
-  <CertificationCenterMembershipsSection @certificationCenterMemberships={{@model.certificationCenterMemberships}} />
+  <CertificationCenters::MembershipsSection @certificationCenterMemberships={{@model.certificationCenterMemberships}} />
 
 </main>
