+++ admin/app/templates/authenticated/certification-centers/new.hbs
@@ -7,7 +7,7 @@
 
 <main class="page-body">
   <section class="page-section">
-    <CertificationCenterForm
+    <CertificationCenters::Form
       @certificationCenter={{@model.certificationCenter}}
       @accreditations={{@model.accreditations}}
       @onSubmit={{this.addCertificationCenter}}
